Move split into its own file
