+++ packages/devtools_app/lib/src/flutter/common_widgets.dart
@@ -2,9 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:math';
-
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widgets/flutter_widgets.dart';
 
@@ -114,165 +111,5 @@ class DefaultTaggedText extends StatelessWidget {
       textAlign: textAlign,
       style: defaultTextStyle,
     );
-  }
-}
-
-/// A widget that takes two children and lays them out along one axis.
-///
-/// The user can customize the amount of space allocated to each child by
-/// dragging a divider between them.
-class Split extends StatefulWidget {
-  /// Builds a [Split] with [Axis.horizontal] direction.
-  ///
-  /// [firstChild] will be placed before [secondChild] in a [Row].
-  const Split.horizontal({
-    Key key,
-    @required Widget firstChild,
-    @required Widget secondChild,
-    double initialFirstFraction,
-  }) : this._(
-          key,
-          Axis.horizontal,
-          firstChild,
-          secondChild,
-          initialFirstFraction,
-        );
-
-  /// Builds a [Split] with [Axis.vertical] direction.
-  ///
-  /// [firstChild] will be placed before [secondChild] in a [Column].
-  const Split.vertical({
-    Key key,
-    @required Widget firstChild,
-    @required Widget secondChild,
-    double initialFirstFraction,
-  }) : this._(
-          key,
-          Axis.vertical,
-          firstChild,
-          secondChild,
-          initialFirstFraction,
-        );
-
-  const Split._(
-    Key key,
-    this.axis,
-    this.firstChild,
-    this.secondChild,
-    double initialFirstFraction,
-  )   : initialFirstFraction = initialFirstFraction ?? 0.5,
-        super(key: key);
-
-  /// The main axis the children will lay out on.
-  ///
-  /// If [Axis.horizontal], the children will be placed in a [Row]
-  /// and they will be horizontally resizable.
-  ///
-  /// If [Axis.vertical], the children will be placed in a [Column]
-  /// and they will be vertically resizable.
-  ///
-  /// Cannot be null.
-  final Axis axis;
-
-  /// The child that will be laid out first along [axis].
-  final Widget firstChild;
-
-  /// The child that will be laid out last along [axis].
-  final Widget secondChild;
-
-  /// The fraction of the layout to allocate to [firstChild].
-  ///
-  /// [secondChild] will receive a fraction of `1 - initialFirstFraction`.
-  final double initialFirstFraction;
-
-  /// The key passed to the divider between [firstChild] and [secondChild].
-  ///
-  /// Visible to grab it in tests.
-  @visibleForTesting
-  Key get dividerKey => Key('$this dividerKey');
-
-  /// The size of the divider between [firstChild] and [secondChild] in
-  /// logical pixels (dp, not px).
-  static const double dividerMainAxisSize = 10.0;
-
-  @override
-  State<StatefulWidget> createState() => _SplitState();
-}
-
-class _SplitState extends State<Split> {
-  double firstFraction;
-  double get secondFraction => 1 - firstFraction;
-  bool get isHorizontal => widget.axis == Axis.horizontal;
-
-  @override
-  void initState() {
-    super.initState();
-    firstFraction = widget.initialFirstFraction;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(builder: _buildLayout);
-  }
-
-  Widget _buildLayout(BuildContext context, BoxConstraints constraints) {
-    final width = constraints.maxWidth;
-    final height = constraints.maxHeight;
-    final axisSize = isHorizontal ? width : height;
-    const halfDivider = Split.dividerMainAxisSize / 2.0;
-    // The fraction of the layout the divider needs to take up from each child.
-    final halfDividerFraction = halfDivider / axisSize;
-
-    final sanitizedFirstFraction =
-        firstFraction.clamp(halfDividerFraction, 1.0 - halfDividerFraction);
-    final sanitizedSecondFraction =
-        secondFraction.clamp(halfDividerFraction, 1.0 - halfDividerFraction);
-
-    final firstSize = axisSize * sanitizedFirstFraction - halfDivider;
-    final secondSize = axisSize * sanitizedSecondFraction - halfDivider;
-
-    void updateSpacing(DragUpdateDetails dragDetails) {
-      final delta = isHorizontal ? dragDetails.delta.dx : dragDetails.delta.dy;
-      final fractionalDelta = delta / axisSize;
-      setState(() {
-        // Update the fraction of space consumed by the children,
-        // being sure not to allocate any of them negative space.
-        firstFraction += fractionalDelta;
-        firstFraction = firstFraction.clamp(0.0, 1.0);
-      });
-    }
-
-    final children = [
-      SizedBox(
-        width: isHorizontal ? firstSize : width,
-        height: isHorizontal ? height : firstSize,
-        child: widget.firstChild,
-      ),
-      GestureDetector(
-        key: widget.dividerKey,
-        onHorizontalDragUpdate: isHorizontal ? updateSpacing : null,
-        onVerticalDragUpdate: isHorizontal ? null : updateSpacing,
-        child: SizedBox(
-          width: isHorizontal ? Split.dividerMainAxisSize : width,
-          height: isHorizontal ? height : Split.dividerMainAxisSize,
-          child: const Center(
-            child: Text(
-              ':::::::',
-              textAlign: TextAlign.center,
-            ),
-          ),
-        ),
-      ),
-      SizedBox(
-        width: isHorizontal ? secondSize : width,
-        height: isHorizontal ? height : secondSize,
-        child: widget.secondChild,
-      ),
-    ];
-    if (widget.axis == Axis.horizontal) {
-      return Row(children: children);
-    } else {
-      return Column(children: children);
-    }
   }
 }
