+++ packages/devtools_app/test/flutter/split_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_app/src/flutter/common_widgets.dart';
+import 'package:devtools_app/src/flutter/split.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
